Crate d3d12

Macros

weak_com_inheritance_chain

Macro that allows generation of an easy to use enum for dealing with many different possible versions of a COM object.

Structs

Binding

CachedPSO

ClearFlags

ComPtr

CommandQueueFlags

D3D12Lib

DescriptorHeapFlags

DescriptorRange

DiscardRegion

DxgiLib

Event

FactoryCreationFlags

HeapDesc

HeapFlags

HeapProperties

IndirectArgument

PipelineStateFlags

PipelineStateSubobject

Subobject of a pipeline stream description

RenderTargetViewDesc

ResourceBarrier

RootParameter

RootSignatureFlags

SampleDesc

Shader

ShaderCompileFlags

StaticSampler

SwapChainPresentFlags

SwapchainDesc

Enums

AlphaMode

CmdListType

CpuPageProperty

DescriptorHeapType

DescriptorRangeType

DxgiAdapter

DxgiFactory

DxgiSwapchain

FeatureLevel

HeapType

MemoryPool

Priority

QueryHeapType

RootSignatureVersion

Scaling

ShaderVisibility

StaticBorderColor

Subobject

SwapEffect

Type Aliases

Adapter1

Adapter2

Adapter3

Adapter4

Blob

BlobResult

CommandAllocator

CommandList

CommandQueue

CommandSignature

CpuDescriptor

D3DResult

Debug

DescriptorHeap

Device

Error

Factory1

Factory2

Factory3

Factory4

Factory5

Factory6

FactoryMedia

Fence

Format

GpuAddress

GpuDescriptor

GraphicsCommandList

HRESULT

Heap

IndexCount

Draw number of indices.

InfoQueue

InstanceCount

Draw number of instances.

NodeMask

PipelineState

QueryHeap

Rect

Resource

RootIndex

Index into the root signature.

RootSignature

Subresource

SwapChain

SwapChain1

SwapChain2

SwapChain3

TextureAddressMode

VertexCount

Draw vertex count.

VertexOffset

Draw vertex base offset.

WorkGroupCount

Number of work groups.